Develop

+++ src/main/java/com/app/advert/portal/controller/CompanyManagementController.java
@@ -1,7 +1,12 @@
 package com.app.advert.portal.controller;
 
 import com.app.advert.portal.dto.CompanyRequestDto;
+import com.app.advert.portal.enums.UserRole;
+import com.app.advert.portal.model.Company;
+import com.app.advert.portal.model.User;
+import com.app.advert.portal.security.SecurityUtils;
 import com.app.advert.portal.service.CompanyService;
+import com.app.advert.portal.service.UserService;
 import io.swagger.annotations.Api;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
@@ -21,13 +26,25 @@ public class CompanyManagementController {
 
     private final CompanyService companyService;
 
+    private final UserService userService;
+
     @PostMapping("/addCompany")
     @Operation(tags = {"Company management"}, description = "Register new company")
     @PreAuthorize("hasAuthority('COMPANY_ADMIN')")
     public ResponseEntity<?> registerNewCompany(@Validated @RequestBody CompanyRequestDto companyDto) {
         try {
             log.info("CompanyController: Register new company");
-            return companyService.saveCompany(companyDto);
+            User user = userService.getBasicUserDataById(SecurityUtils.getLoggedUserId());
+
+            if (user.getCompanyId() != null || user.getRoles().stream().noneMatch(role -> role.getName().equals(UserRole.COMPANY_ADMIN.name()))) {
+                return ResponseEntity.badRequest().body("User has already a company or hasn't access to create a company");
+            }
+            Company company = companyService.saveCompany(companyDto, user.getId());
+            if (company == null) {
+                return ResponseEntity.unprocessableEntity().body("Company with name " + companyDto.getName() + " already exists");
+            }
+
+            return ResponseEntity.ok().body(company);
         } catch (Exception e) {
             return ResponseEntity.internalServerError().body(e);
         }
@@ -39,7 +56,13 @@ public class CompanyManagementController {
     public ResponseEntity<?> deleteCompany(@PathVariable("id") Long companyId) {
         try {
             log.info("CompanyController: Delete company with id: " + companyId);
-            return companyService.deleteCompany(companyId);
+            User user = userService.getBasicUserDataById(SecurityUtils.getLoggedUserId());
+
+            if (!user.getCompanyId().equals(companyId) || user.getRoles().stream().noneMatch(role -> role.getName().equals(UserRole.COMPANY_USER.name()))) {
+                return ResponseEntity.badRequest().body("No access to resource");
+            }
+            companyService.deleteCompany(companyId);
+            return ResponseEntity.ok().build();
         } catch (Exception e) {
             return ResponseEntity.internalServerError().body(e);
         }
@@ -51,7 +74,16 @@ public class CompanyManagementController {
     public ResponseEntity<?> updateCompany(@Validated @RequestBody CompanyRequestDto companyDto) {
         try {
             log.info("CompanyController: Update company with id: " + companyDto.getId());
-            return companyService.updateCompany(companyDto);
+            User user = userService.getBasicUserDataById(SecurityUtils.getLoggedUserId());
+
+            if (user.getCompanyId() != null || user.getRoles().stream().noneMatch(role -> role.getName().equals(UserRole.COMPANY_ADMIN.name()))) {
+                return ResponseEntity.badRequest().body("User has already a company or hasn't access to create a company");
+            }
+            if (companyService.getBasicCompanyDataByName(companyDto.getName()) == null) {
+                return ResponseEntity.unprocessableEntity().body("Company with name " + companyDto.getName() + " not exists");
+            }
+
+            return ResponseEntity.ok().body(companyService.updateCompany(companyDto));
         } catch (Exception e) {
             return ResponseEntity.internalServerError().body(e);
         }
