+++ src/main/java/com/app/advert/portal/service/AmazonS3ClientService.java
@@ -48,11 +48,14 @@ public class AmazonS3ClientService {
         return s3Client.getUrl(bucketName, fileKey).toString();
     }
 
-    public String getFilePresignedFileUrl(String bucketName, String fileKey) {
+    public String getFileAssignedFileUrl(String bucketName, String fileKey) {
         return s3Client.generatePresignedUrl(bucketName, fileKey, new Date(System.currentTimeMillis() + 3600000)).toString();
     }
 
     public void deleteFile(String bucketName, String fileKey) {
+        if (!checkIfBucketExists(bucketName)) {
+            return;
+        }
         s3Client.deleteObject(bucketName, fileKey);
     }
 }
